# proto/gen/openapiv2/api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "WalicaCloneApi"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/event/add": {
      "post": {
        "operationId": "WalicaCloneApi_CreateEvent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateEventReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateEventRequest"
            }
          }
        ],
        "tags": [
          "WalicaCloneApi"
        ]
      }
    },
    "/v1/event/info": {
      "post": {
        "operationId": "WalicaCloneApi_ReadInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiReadInfoReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiReadInfoRequest"
            }
          }
        ],
        "tags": [
          "WalicaCloneApi"
        ]
      }
    },
    "/v1/payment/add": {
      "post": {
        "operationId": "WalicaCloneApi_AddPayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiAddPaymentReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiAddPaymentRequest"
            }
          }
        ],
        "tags": [
          "WalicaCloneApi"
        ]
      }
    },
    "/v1/payment/update": {
      "post": {
        "operationId": "WalicaCloneApi_UpdatePayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUpdatePaymentReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUpdatePaymentRequest"
            }
          }
        ],
        "tags": [
          "WalicaCloneApi"
        ]
      }
    },
    "/v1/user/all": {
      "post": {
        "operationId": "WalicaCloneApi_ReadAllUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiReadAllUsersReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiReadAllUsersRequest"
            }
          }
        ],
        "tags": [
          "WalicaCloneApi"
        ]
      }
    }
  },
  "definitions": {
    "apiAddPaymentReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "apiAddPaymentRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "payerId": {
          "type": "string"
        },
        "eventId": {
          "type": "string"
        },
        "payeeIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiCreateEventReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "apiCreateEventRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiExchange": {
      "type": "object",
      "properties": {
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "payer": {
          "$ref": "#/definitions/apiUser"
        },
        "payee": {
          "$ref": "#/definitions/apiUser"
        }
      }
    },
    "apiPayment": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "payer": {
          "$ref": "#/definitions/apiUser"
        },
        "payees": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiUser"
          }
        },
        "id": {
          "type": "string"
        }
      }
    },
    "apiPaymentSummary": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/apiUser"
        },
        "totalExpense": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apiReadAllUsersReply": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiUser"
          }
        }
      }
    },
    "apiReadAllUsersRequest": {
      "type": "object",
      "properties": {
        "eventId": {
          "type": "string"
        }
      }
    },
    "apiReadInfoReply": {
      "type": "object",
      "properties": {
        "payments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiPayment"
          }
        },
        "exchanges": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiExchange"
          }
        },
        "summaries": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiPaymentSummary"
          }
        },
        "totalExpense": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apiReadInfoRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "apiUpdatePaymentReply": {
      "type": "object",
      "properties": {
        "paymentId": {
          "type": "string"
        }
      }
    },
    "apiUpdatePaymentRequest": {
      "type": "object",
      "properties": {
        "paymentId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "payerId": {
          "type": "string"
        },
        "payeeIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}